refactor sync_to_vllm, sync_from_vllm
