+++ vllm_rbln/utils/optimum/cache_blocks.py
@@ -50,107 +50,3 @@ def get_block_ratio(vllm_config: VllmConfig) -> int:
     else:
         blk_ratio = 1
     return blk_ratio
-
-
-def apply_prefix_caching_block_size(
-    vllm_config: VllmConfig, kvcache_block_size: int, prefill_chunk_size: int
-) -> None:
-    assert prefill_chunk_size is not None, (
-        "prefill_chunk_size must be specified in rbln_config.json"
-    )
-    # If user set prefix_block_size in additional_config, use it.
-    # Otherwise, set it to prefill_chunk_size.
-    prefix_block_size = vllm_config.additional_config.get("prefix_block_size", None)
-    if prefix_block_size is None:
-        prefix_block_size = prefill_chunk_size
-        logger.debug(
-            "Prefix block size is set to %s based on prefill_chunk_size",
-            prefix_block_size,
-        )
-    else:
-        if prefix_block_size % prefill_chunk_size != 0:
-            raise ValueError(
-                "prefix_block_size ({}) is not divisible "
-                "by prefill_chunk_size ({}). "
-                "Please check the value of prefill_chunk_size "
-                "in rbln_config.json".format(prefix_block_size, prefill_chunk_size)
-            )
-        if prefix_block_size > kvcache_block_size:
-            raise ValueError(
-                "prefix_block_size ({}) is greater than "
-                "kvcache_block_size ({}). "
-                "Please check the value of kvcache_block_size "
-                "in rbln_config.json".format(prefix_block_size, kvcache_block_size)
-            )
-        logger.debug(
-            "Prefix block size is set to %s based on additional_config",
-            prefix_block_size,
-        )
-    if kvcache_block_size % prefix_block_size != 0:
-        raise ValueError(
-            "kvcache_block_size ({}) is not divisible "
-            "by prefix_block_size ({}). "
-            "Please check the value of prefix_block_size in rbln_config.json".format(
-                kvcache_block_size, prefix_block_size
-            )
-        )
-    vllm_config.cache_config.block_size = prefix_block_size
-    vllm_config.additional_config["attn_block_size"] = kvcache_block_size
-
-
-def sync_cache_block_size(
-    vllm_config: VllmConfig, kvcache_block_size: int, prefill_chunk_size: int
-) -> None:
-    # vLLM's CacheConfig has a pydantic validator (_apply_block_size_default)
-    # that replaces block_size with DEFAULT_BLOCK_SIZE (16) when
-    # user_specified_block_size is False.  This validator re-runs whenever
-    # VllmConfig is deserialised in a subprocess (e.g. EngineCore), which
-    # would silently overwrite the RBLN-derived block_size we set below.
-    # Mark it as user-specified so the validator preserves our value.
-    vllm_config.cache_config.user_specified_block_size = True
-    if vllm_config.cache_config.enable_prefix_caching:
-        apply_prefix_caching_block_size(
-            vllm_config, kvcache_block_size, prefill_chunk_size
-        )
-    else:
-        if vllm_config.cache_config.block_size != kvcache_block_size:
-            logger.info(
-                "Updating model_cache_config.block_size from %s to %s "
-                "based on rbln_config.json",
-                vllm_config.cache_config.block_size,
-                kvcache_block_size,
-            )
-            vllm_config.cache_config.block_size = kvcache_block_size
-
-
-def sync_num_blocks(vllm_config: VllmConfig, num_blocks: int) -> None:
-    # This function is called twice during startup: once in the main process
-    # when we first read rbln_config.json, and once in each subprocess when
-    # the VllmConfig is deserialized. We only want to perform the num_blocks
-    # calculation and update once, in the main process, to avoid redundant
-    # calculations and potential inconsistencies. We use an additional_config
-    # flag to track whether we have already synced num_blocks.
-    if vllm_config.additional_config.get("num_blocks_synced", False):
-        logger.debug(
-            "num_blocks already synced to %s, skipping...",
-            vllm_config.cache_config.num_gpu_blocks,
-        )
-        return
-    # num_blocks is determined by rbln_config or overridden by user.
-    if vllm_config.cache_config.num_gpu_blocks_override is not None:
-        num_blocks = vllm_config.cache_config.num_gpu_blocks_override
-        # This is kept for optimum based num blocks
-        # not considering ob-ib logic for prefix caching
-        vllm_config.additional_config["num_blocks_override"] = num_blocks
-    blk_ratio = get_block_ratio(vllm_config)
-
-    if is_full_block_available(num_blocks, vllm_config):
-        adjusted_num_blocks = num_blocks * blk_ratio + 1
-    else:
-        adjusted_num_blocks = (num_blocks - 1) * blk_ratio + 1
-
-    vllm_config.cache_config.num_gpu_blocks = adjusted_num_blocks
-
-    if vllm_config.cache_config.num_gpu_blocks_override is not None:
-        vllm_config.cache_config.num_gpu_blocks_override = adjusted_num_blocks
-    vllm_config.additional_config["num_blocks_synced"] = True
