+++ vllm_rbln/utils/optimum/configuration.py
@@ -26,27 +26,25 @@ else:
 
 import vllm_rbln.rbln_envs as envs
 from vllm_rbln.logger import init_logger
-from vllm_rbln.utils.optimum.cache_blocks import (
-    sync_cache_block_size,
-    sync_num_blocks,
-)
 from vllm_rbln.utils.optimum.configuration_helper import (
     keep_only_device_keys,
     strip_runtime_only_keys,
 )
-from vllm_rbln.utils.optimum.rbln_params import (
+from vllm_rbln.utils.optimum.converter import sync_from_vllm, sync_to_vllm
+from vllm_rbln.utils.optimum.params import (
     RBLNParams,
     get_rbln_config,
 )
 from vllm_rbln.utils.optimum.registry import (
     get_rbln_model_info,
-    is_enc_dec_arch,
-    is_generation_arch,
-    is_multi_modal,
-    is_pooling_arch,
 )
 
 logger = init_logger(__name__)
+
+
+# ---------------------------------------------------------------------------
+# Exported utilities / shared helpers
+# ---------------------------------------------------------------------------
 
 
 def get_attn_block_size(vllm_config: VllmConfig) -> int:
@@ -55,6 +53,16 @@ def get_attn_block_size(vllm_config: VllmConfig) -> int:
     else:
         block_size = vllm_config.cache_config.block_size
     return block_size
+
+
+def is_qwen3_pooling(
+    vllm_config: VllmConfig,
+) -> bool:
+    _, model_cls_name = get_rbln_model_info(vllm_config.model_config)
+    return (
+        model_cls_name in ["RBLNQwen3ForCausalLM"]
+        and vllm_config.model_config.runner_type == "pooling"
+    )
 
 
 def generate_model_path_name(
@@ -85,175 +93,6 @@ def generate_model_path_name(
 
     sanitized_name = model_name.replace("/", "_").replace(":", "_")
     return f"{sanitized_name}_{config_hash}"
-
-
-def is_qwen3_pooling(
-    vllm_config: VllmConfig,
-) -> bool:
-    _, model_cls_name = get_rbln_model_info(vllm_config.model_config)
-    return (
-        model_cls_name in ["RBLNQwen3ForCausalLM"]
-        and vllm_config.model_config.runner_type == "pooling"
-    )
-
-
-def update_max_num_batched_tokens(vllm_config: VllmConfig, max_model_len: int) -> None:
-    """
-    Update the max_num_batched_tokens in the vLLM configuration based on the model's
-    maximum length and architecture.
-
-    For encoder-decoder multimodal models (e.g. Whisper), max_num_batched_tokens
-    must be at least max_source_positions so that vllm's MultiModalBudget
-    validation passes (it requires max_tokens_per_mm_item <= max_num_batched_tokens
-    when chunked MM input is disabled).
-    """
-    target_max_num_batched_tokens = max_model_len
-    hf_config = vllm_config.model_config.hf_config
-    if is_enc_dec_arch(hf_config):
-        max_source_positions = getattr(hf_config, "max_source_positions", 0)
-        if max_source_positions > target_max_num_batched_tokens:
-            target_max_num_batched_tokens = max_source_positions
-            logger.info(
-                "Encoder-decoder model detected: setting max_num_batched_tokens "
-                "to %d (max_source_positions) instead of %d (max_model_len)",
-                max_source_positions,
-                max_model_len,
-            )
-
-    cur = vllm_config.scheduler_config.max_num_batched_tokens
-    if cur != target_max_num_batched_tokens:
-        logger.info(
-            "Updating scheduler_config.max_num_batched_tokens "
-            "from %s to %d based on rbln_config.json",
-            cur,
-            target_max_num_batched_tokens,
-        )
-        vllm_config.scheduler_config.max_num_batched_tokens = (
-            target_max_num_batched_tokens
-        )
-
-
-def sync_vllm_from_rbln_config(
-    vllm_config: VllmConfig,
-    params: RBLNParams,
-) -> None:
-    assert params.num_blocks is not None, (
-        "num_blocks must be specified in rbln_config.json"
-    )
-    assert params.batch_size is not None, (
-        "batch_size must be specified in rbln_config.json"
-    )
-    assert params.max_seq_len is not None, (
-        "max_seq_len must be specified in rbln_config.json"
-    )
-    assert params.kvcache_block_size is not None, (
-        "kvcache_block_size must be specified in rbln_config.json"
-    )
-
-    if vllm_config.scheduler_config.max_num_seqs != params.batch_size:
-        logger.info(
-            "Updating scheduler_config.max_num_seqs from %s to %s "
-            "based on rbln_config.json",
-            vllm_config.scheduler_config.max_num_seqs,
-            params.batch_size,
-        )
-        vllm_config.scheduler_config.max_num_seqs = params.batch_size
-
-    update_max_num_batched_tokens(vllm_config, params.max_seq_len)
-
-    if vllm_config.model_config.max_model_len != params.max_seq_len:
-        logger.info(
-            "Updating model_config.max_model_len "
-            "from %s to %s "
-            "based on rbln_config.json",
-            vllm_config.model_config.max_model_len,
-            params.max_seq_len,
-        )
-        vllm_config.model_config.max_model_len = params.max_seq_len
-
-    # Set block_size in cache_config based on rbln_config.json
-    sync_cache_block_size(
-        vllm_config, params.kvcache_block_size, params.prefill_chunk_size
-    )
-    # Set num_blocks in cache_config based on rbln_config.json
-    sync_num_blocks(vllm_config, params.num_blocks)
-    envs.VLLM_RBLN_TP_SIZE = params.tensor_parallel_size
-
-
-def _resolve_compile_block_size(vllm_config: VllmConfig) -> None:
-    """Default / validate ``cache_config.block_size`` prior to compilation.
-
-    When the user did not pin block_size, it defaults to ``max_model_len``.
-    When the user did pin it, enforce the per-architecture constraints that
-    the compiler expects.
-    """
-    cache_config = vllm_config.cache_config
-    hf_config = vllm_config.model_config.hf_config
-    max_model_len = vllm_config.model_config.max_model_len
-
-    if not cache_config.user_specified_block_size:
-        cache_config.block_size = max_model_len
-        return
-
-    if is_multi_modal(hf_config) or is_generation_arch(hf_config):
-        assert cache_config.block_size >= 4096, (
-            "block_size must be at least 4096 for compilation."
-        )
-    if is_pooling_arch(hf_config):
-        assert cache_config.block_size == max_model_len, (
-            "For pooling models, block_size must be equal to max_model_len."
-        )
-
-
-def prepare_vllm_for_compile(vllm_config: VllmConfig) -> None:
-    """Stage ``vllm_config`` with user-provided compile parameters.
-
-    Mirrors any explicit fields from ``additional_config['rbln_config']`` into
-    the matching ``vllm_config`` slots (these are what the compile step will
-    feed to optimum-rbln), then resolves ``block_size`` and derives
-    ``max_num_batched_tokens``. ``num_blocks`` is populated post-compile via
-    :func:`sync_num_blocks`.
-    """
-    rbln_config = vllm_config.additional_config.get("rbln_config", {})
-    params = RBLNParams.from_rbln_config(vllm_config, rbln_config)
-
-    if params.batch_size is not None:
-        logger.info(
-            "Setting max_num_seqs to %d based on rbln_config in additional_config",
-            params.batch_size,
-        )
-        vllm_config.scheduler_config.max_num_seqs = params.batch_size
-    if params.max_seq_len is not None:
-        logger.info(
-            "Setting max_model_len to %d based on rbln_config in additional_config",
-            params.max_seq_len,
-        )
-        vllm_config.model_config.max_model_len = params.max_seq_len
-    if params.kvcache_block_size is not None:
-        logger.info(
-            "Setting block_size to %d based on rbln_config in additional_config",
-            params.kvcache_block_size,
-        )
-        vllm_config.cache_config.block_size = params.kvcache_block_size
-        vllm_config.cache_config.user_specified_block_size = params.kvcache_block_size
-
-    _resolve_compile_block_size(vllm_config)
-    sync_cache_block_size(
-        vllm_config, vllm_config.cache_config.block_size, prefill_chunk_size=128
-    )
-
-    # max_num_batched_tokens must fit both a full-length prefill and a full
-    # batch dispatch; update_max_num_batched_tokens layers the enc-dec
-    # max_source_positions constraint on top.
-    vllm_config.scheduler_config.max_num_batched_tokens = max(
-        vllm_config.model_config.max_model_len,
-        vllm_config.scheduler_config.max_num_seqs,
-    )
-    update_max_num_batched_tokens(
-        vllm_config, vllm_config.scheduler_config.max_num_batched_tokens
-    )
-
-    logger.info("Prepared vLLM config for compilation: %s", vllm_config)
 
 
 def _resolve_rbln_config(vllm_config: VllmConfig) -> dict | None:
@@ -293,7 +132,7 @@ def sync_with_rbln_config(vllm_config: VllmConfig) -> None:
     """
     rbln_config = _resolve_rbln_config(vllm_config)
     if rbln_config is None:
-        prepare_vllm_for_compile(vllm_config)
+        sync_from_vllm(vllm_config)
         return
 
     # Pre-compiled (or cache-hit): the compiled artefact is the source of
@@ -303,4 +142,4 @@ def sync_with_rbln_config(vllm_config: VllmConfig) -> None:
         vllm_config.additional_config.get("rbln_config", {})
     )
     params = RBLNParams.from_rbln_config(vllm_config, rbln_config)
-    sync_vllm_from_rbln_config(vllm_config, params)
+    sync_to_vllm(vllm_config, params)
