+++ vllm_rbln/utils/optimum/converter/__init__.py
@@ -1,0 +1,8 @@
+from .optimum_to_vllm import sync_to_vllm, sync_num_blocks
+from .vllm_to_optimum import sync_from_vllm
+
+__all__ = [
+    "sync_to_vllm",
+    "sync_from_vllm",
+    "sync_num_blocks",
+]
